Move checklist to module folder
