+++ src/modules/Checklist/CheckItems/CheckItems.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import styled from 'styled-components';
-import { CheckItem as CheckItemType } from '../../typings/apiTypes';
+import { CheckItem as CheckItemType } from '../../../typings/apiTypes';
 import CheckItem from './CheckItem/CheckItem';
 import CheckHeader from './CheckHeader';
-import { ProcosysApiService } from '../../services/procosysApi';
+import { ProcosysApiService } from '../../../services/procosysApi';
 
 const CheckItemsWrapper = styled.div`
     display: flex;
