+++ src/modules/Checklist/CheckItems/CustomCheckItems.tsx
@@ -2,13 +2,13 @@ import { Button, Checkbox, Scrim, TextField } from '@equinor/eds-core-react';
 import axios from 'axios';
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import EdsIcon from '../../components/icons/EdsIcon';
-import { CustomCheckItem } from '../../typings/apiTypes';
-import { ProcosysApiService } from '../../services/procosysApi';
-import { COLORS, SHADOW } from '../../style/GlobalStyles';
-import updateCustomCheck from '../../utils/updateCustomCheck';
+import EdsIcon from '../../../components/icons/EdsIcon';
+import { CustomCheckItem } from '../../../typings/apiTypes';
+import { ProcosysApiService } from '../../../services/procosysApi';
+import { COLORS, SHADOW } from '../../../style/GlobalStyles';
+import updateCustomCheck from '../../../utils/updateCustomCheck';
 import CheckHeader from './CheckHeader';
-import { AsyncStatus } from '../../typings/enums';
+import { AsyncStatus } from '../../../typings/enums';
 
 const CustomCheckItemWrapper = styled.div`
     display: flex;
