+++ src/modules/Checklist/ChecklistMultiSignOrVerify.tsx
@@ -1,10 +1,10 @@
 import { Button, Checkbox, Scrim } from '@equinor/eds-core-react';
 import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
-import { ItemToMultiSignOrVerify } from '../typings/apiTypes';
-import { ProcosysApiService } from '../services/procosysApi';
-import { COLORS } from '../style/GlobalStyles';
-import { AsyncStatus } from '../typings/enums';
+import { ItemToMultiSignOrVerify } from '../../typings/apiTypes';
+import { ProcosysApiService } from '../../services/procosysApi';
+import { COLORS } from '../../style/GlobalStyles';
+import { AsyncStatus } from '../../typings/enums';
 
 export const MultiSignVerifyContainer = styled.div`
     border-radius: 8px;
